add support manual controls
This will add manual controlls to the app with flows
!! untested still !!

I have chosen to control the set manual controls with flows only because
of 2 reasons
1:
since there is a maximum of 6 commands to the valve that it can handlt
when it wakes up
and since this needs to be in manual mode, it will already use up 2
commands
in mobile cards there is a huge possibility that you either set
accidentally or can't find the good value in the first go then every
stop in between will also keep that in the queue
2:
there was no good layout to add the slider to the mobile card, and it
still let it look good.

there are flow cards for when it is changed (re-pair needed), action
cards to send a new value (no re-pair needed) and the global token with
the current value. (re-pair needed)