remove debug tekst, add chartLine, add small timeout
remove debug tekst, working now as intended.

add a chart type, that should activate insights (re-pair might be
needed)

added a small timeout if the mode needs to change to make sure the mode
is changed before the (new) level is set.
in theorie this is not needed, but if for some reason something came
inbetween the sending of the mode, this will make sure it is set
beforehand.